# docs/openapi/relay.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "AI模型接口",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/v1/assets": {
      "post": {
        "summary": "创建素材",
        "deprecated": false,
        "operationId": "createAsset",
        "tags": [
          "素材"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAssetRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "素材已在所选 Seedance 渠道创建",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Asset"
                }
              }
            }
          },
          "400": {
            "description": "素材请求无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "素材、模型或渠道状态冲突"
          },
          "502": {
            "description": "素材上游操作失败"
          },
          "503": {
            "description": "素材上游不可用"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "get": {
        "summary": "列出素材",
        "deprecated": false,
        "operationId": "listAssets",
        "tags": [
          "素材"
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "asset_kind",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "media_type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "素材列表",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetListResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/v1/assets/{asset_id}": {
      "get": {
        "summary": "查询素材",
        "deprecated": false,
        "operationId": "getAsset",
        "tags": [
          "素材"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/AssetId"
          }
        ],
        "responses": {
          "200": {
            "description": "素材详情",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Asset"
                }
              }
            }
          },
          "404": {
            "description": "素材不存在"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "patch": {
        "summary": "更新素材名称",
        "deprecated": false,
        "operationId": "updateAsset",
        "tags": [
          "素材"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/AssetId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAssetRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新后的素材",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Asset"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "delete": {
        "summary": "删除素材",
        "deprecated": false,
        "operationId": "deleteAsset",
        "tags": [
          "素材"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/AssetId"
          }
        ],
        "responses": {
          "204": {
            "description": "素材已删除"
          },
          "409": {
            "description": "素材状态或绑定不允许删除"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/v1/asset-groups": {
      "post": {
        "summary": "创建素材组或真人认证会话",
        "deprecated": false,
        "operationId": "createAssetGroup",
        "tags": [
          "素材"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAssetGroupRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "素材组已创建；真人组包含上游认证地址",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetGroup"
                }
              }
            }
          },
          "400": {
            "description": "素材组请求无效"
          },
          "502": {
            "description": "素材上游操作失败"
          },
          "503": {
            "description": "素材上游不可用"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "get": {
        "summary": "列出素材组",
        "deprecated": false,
        "operationId": "listAssetGroups",
        "tags": [
          "素材"
        ],
        "responses": {
          "200": {
            "description": "素材组列表",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetGroupListResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/v1/asset-groups/{group_id}": {
      "get": {
        "summary": "查询素材组或真人认证状态",
        "deprecated": false,
        "operationId": "getAssetGroup",
        "tags": [
          "素材"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/AssetGroupId"
          }
        ],
        "responses": {
          "200": {
            "description": "素材组详情",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetGroup"
                }
              }
            }
          },
          "404": {
            "description": "素材组不存在"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "delete": {
        "summary": "删除素材组",
        "deprecated": false,
        "operationId": "deleteAssetGroup",
        "tags": [
          "素材"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/AssetGroupId"
          }
        ],
        "responses": {
          "204": {
            "description": "素材组已删除"
          },
          "409": {
            "description": "素材组仍包含素材或当前状态不允许删除"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string",
                "description": "错误信息"
              },
              "type": {
                "type": "string",
                "description": "错误类型"
              },
              "param": {
                "type": "string",
                "description": "相关参数",
                "nullable": true
              },
              "code": {
                "type": "string",
                "description": "错误代码",
                "nullable": true
              }
            }
          }
        }
      },
...
      "AssetSource": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "type",
          "url",
          "expires_at"
        ],
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "url"
            ]
          },
          "url": {
            "type": "string",
            "format": "uri"
          },
          "expires_at": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "CreateAssetRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "name",
          "asset_kind",
          "media_type",
          "model",
          "source"
        ],
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 64
          },
          "asset_kind": {
            "type": "string",
            "enum": [
              "general",
              "real_person"
            ]
          },
          "media_type": {
            "type": "string",
            "enum": [
              "image",
              "video",
              "audio"
            ]
          },
          "model": {
            "type": "string"
          },
          "asset_group_id": {
            "type": "string",
            "description": "可选 astgrp_*；真人素材必须属于已认证真人组"
          },
          "source": {
            "$ref": "#/components/schemas/AssetSource"
          }
        }
      },
      "UpdateAssetRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
      "Asset": {
        "type": "object",
        "required": [
          "id",
          "name",
          "asset_kind",
          "media_type",
          "model",
          "status",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^ast_"
          },
          "name": {
            "type": "string"
          },
          "asset_kind": {
            "type": "string",
            "enum": [
              "general",
              "real_person"
            ]
          },
          "media_type": {
            "type": "string",
            "enum": [
              "image",
              "video",
              "audio"
            ]
          },
          "model": {
            "type": "string"
          },
          "asset_group_id": {
            "type": "string",
            "pattern": "^astgrp_"
          },
          "status": {
            "type": "string",
            "enum": [
              "processing",
              "ready",
              "failed",
              "deleted"
            ]
          },
          "error_code": {
            "type": "string"
          },
          "error": {
            "type": "string"
          },
          "created_at": {
            "type": "integer",
            "format": "int64"
          },
          "updated_at": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "AssetListResponse": {
        "type": "object",
        "required": [
          "data",
          "total",
          "page",
          "page_size"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Asset"
            }
          },
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "page": {
            "type": "integer"
          },
          "page_size": {
            "type": "integer"
          }
        }
      },
      "CreateAssetGroupRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "name",
          "group_kind",
          "model"
        ],
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 64
          },
          "description": {
            "type": "string",
            "maxLength": 300
          },
          "group_kind": {
            "type": "string",
            "enum": [
              "general",
              "real_person"
            ]
          },
          "model": {
            "type": "string"
          },
          "redirect_url": {
            "type": "string",
            "format": "uri",
            "description": "可选；直接传给上游真人认证服务"
          }
        }
      },
      "AssetGroup": {
        "type": "object",
        "required": [
          "id",
          "name",
          "group_kind",
          "model",
          "status",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^astgrp_"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "group_kind": {
            "type": "string",
            "enum": [
              "general",
              "real_person"
            ]
          },
          "model": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "processing",
              "ready",
              "failed",
              "deleted"
            ]
          },
          "verification_url": {
            "type": "string",
            "format": "uri",
            "description": "上游真人认证地址；仅在有效期内返回"
          },
          "expires_at": {
            "type": "integer",
            "format": "int64"
          },
          "error_code": {
            "type": "string"
          },
          "error": {
            "type": "string"
          },
          "created_at": {
            "type": "integer",
            "format": "int64"
          },
          "updated_at": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "AssetGroupListResponse": {
        "type": "object",
        "required": [
          "data",
          "total",
          "page",
          "page_size"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AssetGroup"
            }
          },
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "page": {
            "type": "integer"
          },
          "page_size": {
            "type": "integer"
          }
        }
      }
    },
    "parameters": {
...
      "AssetId": {
        "name": "asset_id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
      "AssetGroupId": {
        "name": "group_id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "pattern": "^astgrp_"
        }
      }
    },
...
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "使用 Bearer Token 认证。\n格式: `Authorization: Bearer sk-xxxxxx`\n"
      },
      "SessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session",
        "description": "已废弃：Gin session 不再用于面板鉴权"
      },
      "AccessToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT or PAT",
        "description": "面板短期 JWT 或 User.AccessToken PAT；不再要求 New-Api-User"
      },
      "NewApiUser": {
        "type": "apiKey",
        "in": "header",
        "name": "New-Api-User",
        "description": "已废弃且不再参与鉴权；PAT 客户端可以移除此请求头"
      },
      "Combination": {
        "group": [
          {
            "id": "AccessToken"
          }
        ],
        "type": "combination"
      },
      "Combination1": {
        "group": [
          {
            "id": "AccessToken"
          }
        ],
        "type": "combination"
      },
      "Combination2": {
...
  },
  "servers": [],
  "security": [
    {
      "BearerAuth": []
    }
  ]
}